Moving shared functions to choropleth utils
